refresh overlay
